WIP: Reorganize code into sub folders
