+++ src/background/background.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8">
     <title>FBTR Background Page</title>
     <script src="/browser-polyfill.min.js"></script>
-    <script src="../background.js" type="module"></script>
+    <script src="background.js" type="module"></script>
 </head>
 <body>
 </body>
