+++ src/content_script/content_script.js
@@ -1,3 +1,3 @@
 
-import(browser.runtime.getURL("src/app.js"))
+import(browser.runtime.getURL("src/content_script/app.js"))
     .catch(e => console.warn(`Failed to load FacebookTrackingRemoval: ${e}`));
